refactor: modular parsers
