+++ webcomponents/src/components/rsvp-player.ts
@@ -3,13 +3,7 @@ import { property, state } from 'lit/decorators.js';
 import './rsvp-settings';
 import './rsvp-controls';
 import './rsvp-fullscreen';
-
-export interface Token {
-  text: string;
-  scopes: string[];
-  markers: string[];
-  extraPause: number;
-}
+import { Token, parseText } from '../parsers/tokenizer';
 
 interface Keybindings {
   playPause: string;
@@ -25,117 +19,6 @@ interface GestureSettings {
   settingsSwipe: boolean;
 }
 
-const OPENERS = ['(', '[', '{', '"', "'"] as const;
-const CLOSERS: Record<string, string> = {
-  ')': '(',
-  ']': '[',
-  '}': '{',
-  '"': '"',
-  "'": "'",
-};
-
-/* eslint-disable max-lines-per-function, sonarjs/cognitive-complexity, complexity */
-export function parseText(text: string): Token[] {
-  const tokens: Token[] = [];
-  const stack: string[] = [];
-  let word = '';
-  let sentenceIndices: number[] = [];
-
-  const pushWord = () => {
-    if (word) {
-      tokens.push({ text: word, scopes: [...stack], markers: [], extraPause: 0 });
-      sentenceIndices.push(tokens.length - 1);
-      word = '';
-    }
-  };
-
-  for (let i = 0; i < text.length;) {
-    const ch = text[i];
-
-    if (text.slice(i, i + 3) === '...') {
-      pushWord();
-      for (let j = 1; j <= 3; j++) {
-        tokens.push({ text: '.'.repeat(j), scopes: [...stack], markers: [], extraPause: 0 });
-        sentenceIndices.push(tokens.length - 1);
-      }
-      i += 3;
-      continue;
-    }
-
-    if ((OPENERS as readonly string[]).includes(ch)) {
-      pushWord();
-      stack.push(ch);
-      i++;
-      continue;
-    }
-
-    if (Object.prototype.hasOwnProperty.call(CLOSERS, ch)) {
-      pushWord();
-      if (stack.at(-1) === CLOSERS[ch]) {
-        stack.pop();
-      }
-      i++;
-      continue;
-    }
-
-    if (/\s/.test(ch)) {
-      pushWord();
-      i++;
-      continue;
-    }
-
-    if (ch === ',') {
-      pushWord();
-      if (tokens.length > 0) {
-        const last = tokens.at(-1)!;
-        last.extraPause += 1;
-      }
-      i++;
-      continue;
-    }
-
-    if (ch === ':' || ch === ';') {
-      pushWord();
-      if (tokens.length > 0) {
-        const last = tokens.at(-1)!;
-        last.extraPause += 1;
-      }
-      i++;
-      continue;
-    }
-
-    if (ch === '.' || ch === '!' || ch === '?') {
-      pushWord();
-      let j = i;
-      const markers: string[] = [];
-      while (j < text.length) {
-        if (text.slice(j, j + 3) === '...') {
-          break;
-        }
-        const c = text[j];
-        if (c === '!' || c === '?') {
-          markers.push(c);
-        } else if (c !== '.') {
-          break;
-        }
-        j++;
-      }
-      const dedup = [...new Set(markers)];
-      for (const idx of sentenceIndices) {
-        tokens[idx].markers = dedup;
-      }
-      sentenceIndices = [];
-      i = j;
-      continue;
-    }
-
-    word += ch;
-    i++;
-  }
-  pushWord();
-  return tokens;
-}
-/* eslint-enable max-lines-per-function, sonarjs/cognitive-complexity */
 
 function formatToken(token: Token): string {
   const closingMap: Record<string, string> = {
